Remove impl_parser: all languages now use shared extract module
Delete src/check/impl_parser.rs — its responsibilities are fully covered:
- Type definitions (ExtractedImpl, ExtractedStruct, etc.) moved to check/mod.rs
- Rust parsing replaced by rust_extractor (shared with extract command)
- Operation extraction handled by extract_fns_generic (now includes 'pub fn')

Removed 14 impl_parser-specific tests (type_to_mult, parse_impl, unit struct
regression tests). Differ tests retained with direct type construction.

Updated CLAUDE.md module listing and all comment references.
